requirements.txt
# Core dependencies
streamlit>=1.31.0
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.22.0

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0

# Dimensionality reduction and TDA
umap-learn>=0.5.5
# giotto-tda==0.1.4  # Commented out due to complex build requirements in Docker

# Alternative TDA libraries (Docker-friendly with pre-compiled wheels)
kmapper>=2.0.0        # Mapper algorithm for topological network visualization
ripser>=0.6.0         # Fast Vietoris-Rips persistent homology
persim>=0.3.0         # Persistence diagram distances and visualizations
scikit-tda>=1.0.0     # Additional TDA utilities
# gudhi>=3.8.0        # Not available for ARM64/Python 3.11 - would need local build

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0

# Linting and formatting
ruff>=0.4.0

# Type checking
mypy>=1.8.0
